third_party/WebKit/LayoutTests/editing/spelling/delete-misspelled-word.html

<!DOCTYPE html>
<html>
<head>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script src=resources/util.js></script>
<title>Testing moving cursor to a misspelled word</title>
</head>
<body>
<div id="src" contenteditable></div>
<script>
var node = document.getElementById('src');

var steps = [
    function() {
        node.focus();
        document.execCommand('InsertText', false, 'z');
        document.execCommand('InsertText', false, 'z');
        document.execCommand('InsertText', false, ' ');
    },
    function() {
        testRunner.execCommand('DeleteBackward');
        testRunner.execCommand('DeleteBackward');
        document.execCommand('InsertText', false, ' ');
    }
];

var assertions = [
    () => assert_true(internals.hasSpellingMarker(document, 0, 2)),
    () => assert_false(internals.hasSpellingMarker(document, 0, 1))
];

runSpellingTest(steps, assertions, 'Blink spellchecks a word again when changing a misspelled word');
</script>
</body>
</html>
